FPU64 fixes, software build system changes, support for imgui
Changes to the software build system. No longer a need for special makefiles for multi-file packages. Makefile has a list of object files for the packages and will compile .c .cpp and .s files to generate the objects for linking. Allows for extra CFLAGS and INCLUDES.

IMGUI demo provided.  Runs the demo window. Test in B&W and RGBA colour. No interaction presently possible.

All P64 files rebuilt using the new build system. New build system backported to RV32.

( RV64 )

ALU.si AMO now takes into account function3 to specify 32/64 bit.

CPU.si use a ++: wait state before executing CSR/FPU/AMO/DIVIDE to allow for removal of output! in cpu_function blocks on the floating point registers. Make changes to conversions to use the new changeprecision unit, combining float2double and double2float. Fix error in detecting IL2FD.

cpu_functionblocks.si remove output! on all registers.

FPU64 simplify MIN/MAX. Simplify sign manipulation.  Simplify overflow/underflow in floatdouble2intlong. Combine float2double and double2float into changeprecision. Simplify newexponent in doroundcombine. Simplify product generation in floatmultiply. Correct xZERO in floatdivide for 32bit, allows IMGUI text to display.

( RV32 )

Backport new build system. Backport new RGBA ARGB pixelblocks, and the DMA support for them.

+++ RV32/SOFTWARE/c/memtest/memtest.c
@@ -37,9 +37,9 @@ int main( void ) {
         while( membyte < (unsigned char *)last ) { *membyte++ = byte; count++; byte = byte ^ count; }
         memword = (unsigned int *)start;
         while( memword < (unsigned int *)lastprint ) {
-            if( ( (int)memword & 0xc ) == 0 ) printw( "0x%7x : ",(int)memword );
+            if( ( (long)memword & 0xc ) == 0 ) printw( "0x%7x : ",(long)memword );
             printw("%08x ",*memword);
-            if( ( (int)memword & 0xc ) == 0xc ) printw( "\n" );
+            if( ( (long)memword & 0xc ) == 0xc ) printw( "\n" );
             memword++;
         }
         membyte = (unsigned char *)start; byte = 0x11; count = 0;
@@ -57,9 +57,9 @@ int main( void ) {
         while( memhalf < (unsigned short *)last ) { *memhalf++ = half; count++; half = half ^ count ^ ( count << 8 ); }
         memword = (unsigned int *)start;
         while( memword < (unsigned int *)lastprint ) {
-            if( ( (int)memword & 0xc ) == 0 ) printw( "0x%7x : ",(int)memword );
+            if( ( (long)memword & 0xc ) == 0 ) printw( "0x%7x : ",(long)memword );
             printw("%08x ",*memword);
-            if( ( (int)memword & 0xc ) == 0xc ) printw( "\n" );
+            if( ( (long)memword & 0xc ) == 0xc ) printw( "\n" );
             memword++;
         }
         memhalf = (unsigned short *)start; half = 0x2211; count = 0;
@@ -77,16 +77,16 @@ int main( void ) {
         while( memword < (unsigned int *)last ) { *memword++ = word; count++; word = word ^ count ^ ( count << 8 ) ^ ( count << 16 ) ^ ( count << 24 ); }
         membyte = (unsigned char *)start;
         while( membyte < (unsigned char *)lastprint ) {
-            if( ( (int)membyte & 0xf ) == 0 ) printw( "0x%7x : ",(int)membyte );
+            if( ( (long)membyte & 0xf ) == 0 ) printw( "0x%7x : ",(long)membyte );
             printw("%02x ",*membyte);
-            if( ( (int)membyte & 0xf ) == 0xf ) printw( "\n" );
+            if( ( (long)membyte & 0xf ) == 0xf ) printw( "\n" );
             membyte++;
         }
         memword = (unsigned int *)start;
         while( memword < (unsigned int *)lastprint ) {
-            if( ( (int)memword & 0xc ) == 0 ) printw( "0x%7x : ",(int)memword );
+            if( ( (long)memword & 0xc ) == 0 ) printw( "0x%7x : ",(long)memword );
             printw("%08x ",*memword);
-            if( ( (int)memword & 0xc ) == 0xc ) printw( "\n" );
+            if( ( (long)memword & 0xc ) == 0xc ) printw( "\n" );
             memword++;
         }
         memword = (unsigned int *)start; word = 0x44332211; count = 0;
